fix: use BigInteger for validation_sessions.organization_id FK
Integration Tests (MySQL) failed with errno 3780:
  "Referencing column 'organization_id' and referenced column 'id' in
  foreign key constraint 'validation_sessions_ibfk_1' are incompatible."

Same root cause as the ml/workflow tables fixed previously:
organizations.id is BIGINT, but validation_sessions.organization_id was
declared as plain Integer with an actual FK constraint (I'd missed the
ForeignKeyConstraint clause a few lines below the column definition when
first reviewing this migration). MySQL enforces matching FK column types,
SQLite doesn't, so this was invisible locally.

Fixed the column type in both
alembic/versions/ab3669d60d26_reconcile_schema_drift_with_current_.py and
validation/models.py (ValidationSession.organization_id) to use BigInteger
/ the existing BigInt type alias, keeping the ORM model and migration in
sync.

Searched the rest of the codebase for the same `Column(Integer,
ForeignKey(...))` pattern against organizations.id/users.id and found no
other occurrences.

+++ alembic/versions/ab3669d60d26_reconcile_schema_drift_with_current_.py
@@ -2000,7 +2000,7 @@ def upgrade() -> None:
         sa.Column("approved_at", sa.DateTime(), nullable=True),
         sa.Column("approved_by", sa.String(length=255), nullable=True),
         sa.Column("approval_comments", sa.Text(), nullable=True),
-        sa.Column("organization_id", sa.Integer(), nullable=True),
+        sa.Column("organization_id", sa.BigInteger(), nullable=True),
         sa.ForeignKeyConstraint(
             ["organization_id"],
             ["organizations.id"],
